+++ zoomimage-compose-coil2-core/src/main/kotlin/com/github/panpf/zoomimage/compose/coil/CoilComposeSubsamplingImageGenerator.kt
@@ -16,7 +16,6 @@
 
 package com.github.panpf.zoomimage.compose.coil
 
-import com.github.panpf.zoomimage.subsampling.ImageSource as ZoomImageImageSource
 import android.content.Context
 import androidx.compose.ui.graphics.painter.Painter
 import coil.ImageLoader
@@ -24,7 +23,7 @@ import coil.request.SuccessResult
 import com.github.panpf.zoomimage.subsampling.SubsamplingImageGenerateResult
 
 /**
- * Convert coil model to [ZoomImageImageSource.Factory]
+ * Convert coil SuccessResult to [SubsamplingImageGenerateResult]
  */
 interface CoilComposeSubsamplingImageGenerator {