+++ zoomimage-compose-coil3-core/src/commonMain/kotlin/com/github/panpf/zoomimage/CoilZoomAsyncImage.kt
@@ -25,7 +25,6 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.geometry.isSpecified
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.DefaultAlpha
@@ -33,34 +32,28 @@ import androidx.compose.ui.graphics.FilterQuality
 import androidx.compose.ui.graphics.drawscope.DrawScope
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.Layout
 import androidx.compose.ui.unit.IntSize
 import coil3.ImageLoader
 import coil3.PlatformContext
+import coil3.compose.AsyncImage
 import coil3.compose.AsyncImagePainter
 import coil3.compose.AsyncImagePainter.State
 import coil3.compose.LocalPlatformContext
 import coil3.request.ImageRequest
 import com.github.panpf.zoomimage.coil.CoilTileImageCache
-import com.github.panpf.zoomimage.compose.coil.internal.AsyncImageState
-import com.github.panpf.zoomimage.compose.coil.internal.ContentPainterElement
-import com.github.panpf.zoomimage.compose.coil.internal.UseMinConstraintsMeasurePolicy
 import com.github.panpf.zoomimage.compose.coil.internal.onStateOf
-import com.github.panpf.zoomimage.compose.coil.internal.previewHandler
-import com.github.panpf.zoomimage.compose.coil.internal.requestOfWithSizeResolver
 import com.github.panpf.zoomimage.compose.coil.internal.transformOf
-import com.github.panpf.zoomimage.compose.coil.internal.validateRequest
 import com.github.panpf.zoomimage.compose.subsampling.subsampling
 import com.github.panpf.zoomimage.compose.zoom.ScrollBarSpec
 import com.github.panpf.zoomimage.compose.zoom.mouseZoom
 import com.github.panpf.zoomimage.compose.zoom.zoom
 import com.github.panpf.zoomimage.compose.zoom.zoomScrollBar
+import com.github.panpf.zoomimage.compose.zoom.zooming
 import com.github.panpf.zoomimage.subsampling.SubsamplingImage
 import com.github.panpf.zoomimage.subsampling.SubsamplingImageGenerateResult
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
-
 
 /**
  * An image component that integrates the Coil image loading framework that zoom and subsampling huge images
@@ -130,8 +123,9 @@ fun CoilZoomAsyncImage(
     onLongPress: ((Offset) -> Unit)? = null,
     onTap: ((Offset) -> Unit)? = null,
 ) = CoilZoomAsyncImage(
-    state = AsyncImageState(model, imageLoader),
+    model = model,
     contentDescription = contentDescription,
+    imageLoader = imageLoader,
     modifier = modifier,
     transform = transformOf(placeholder, error, fallback),
     onState = onStateOf(onLoading, onSuccess, onError),
@@ -189,44 +183,10 @@ fun CoilZoomAsyncImage(
  * @see com.github.panpf.zoomimage.compose.coil3.core.test.CoilZoomAsyncImageTest.testCoilZoomAsyncImage2
  */
 @Composable
-@NonRestartableComposable
 fun CoilZoomAsyncImage(
     model: Any?,
     contentDescription: String?,
     imageLoader: ImageLoader,
-    modifier: Modifier = Modifier,
-    transform: (State) -> State = AsyncImagePainter.DefaultTransform,
-    onState: ((State) -> Unit)? = null,
-    alignment: Alignment = Alignment.Center,
-    contentScale: ContentScale = ContentScale.Fit,
-    alpha: Float = DefaultAlpha,
-    colorFilter: ColorFilter? = null,
-    filterQuality: FilterQuality = DrawScope.DefaultFilterQuality,
-    zoomState: CoilZoomState = rememberCoilZoomState(),
-    scrollBar: ScrollBarSpec? = ScrollBarSpec.Default,
-    onLongPress: ((Offset) -> Unit)? = null,
-    onTap: ((Offset) -> Unit)? = null,
-) = CoilZoomAsyncImage(
-    state = AsyncImageState(model, imageLoader),
-    contentDescription = contentDescription,
-    modifier = modifier,
-    transform = transform,
-    onState = onState,
-    alignment = alignment,
-    contentScale = contentScale,
-    alpha = alpha,
-    colorFilter = colorFilter,
-    filterQuality = filterQuality,
-    zoomState = zoomState,
-    scrollBar = scrollBar,
-    onLongPress = onLongPress,
-    onTap = onTap,
-)
-
-@Composable
-private fun CoilZoomAsyncImage(
-    state: AsyncImageState,
-    contentDescription: String?,
     modifier: Modifier = Modifier,
     transform: (State) -> State = AsyncImagePainter.DefaultTransform,
     onState: ((State) -> Unit)? = null,
@@ -244,44 +204,51 @@ private fun CoilZoomAsyncImage(
     zoomState.zoomable.alignment = alignment
 
     LaunchedEffect(zoomState.subsampling) {
-        zoomState.subsampling.tileImageCache = CoilTileImageCache(state.imageLoader)
+        zoomState.subsampling.tileImageCache = CoilTileImageCache(imageLoader)
     }
 
     // moseZoom directly acts on ZoomAsyncImage, causing the zoom center to be abnormal.
     Box(modifier = modifier.mouseZoom(zoomState.zoomable)) {
         val context = LocalPlatformContext.current
         val coroutineScope = rememberCoroutineScope()
-        BaseCoilZoomAsyncImage(
-            state = state,
+        AsyncImage(
+            model = model,
             contentDescription = contentDescription,
+            imageLoader = imageLoader,
             transform = transform,
-            onState = { loadState ->
-                onState(
-                    context = context,
-                    coroutineScope = coroutineScope,
-                    imageLoader = state.imageLoader,
-                    zoomState = zoomState,
-                    model = state.model,
-                    loadState = loadState
-                )
-                onState?.invoke(loadState)
-            },
             contentScale = contentScale,
             alignment = alignment,
             alpha = alpha,
             colorFilter = colorFilter,
             filterQuality = filterQuality,
             clipToBounds = false,
-            keepContentNoneStartOnDraw = true,
             modifier = Modifier
                 .matchParentSize()
                 .zoom(
                     zoomable = zoomState.zoomable,
                     userSetupContentSize = true,
+                    restoreContentToNoneLeftTopFirst = true,
                     onLongPress = onLongPress,
                     onTap = onTap
+                ),
+            onState = { loadState ->
+                onState(
+                    context = context,
+                    coroutineScope = coroutineScope,
+                    imageLoader = imageLoader,
+                    zoomState = zoomState,
+                    model = model,
+                    loadState = loadState
                 )
-                .subsampling(zoomState.zoomable, zoomState.subsampling),
+                onState?.invoke(loadState)
+            },
+        )
+
+        Box(
+            Modifier
+                .matchParentSize()
+                .zooming(zoomable = zoomState.zoomable, restoreContentToNoneLeftTopFirst = false)
+                .subsampling(zoomState.zoomable, zoomState.subsampling)
         )
 
         if (scrollBar != null) {
@@ -294,55 +261,6 @@ private fun CoilZoomAsyncImage(
     }
 }
 
-/**
- * Copy from coil3.compose.AsyncImage()
- *
- * Add keepContentNoneStartOnDraw param
- */
-@Composable
-private fun BaseCoilZoomAsyncImage(
-    state: AsyncImageState,
-    contentDescription: String?,
-    modifier: Modifier,
-    transform: (State) -> State,
-    onState: ((State) -> Unit)?,
-    alignment: Alignment,
-    contentScale: ContentScale,
-    alpha: Float,
-    colorFilter: ColorFilter?,
-    filterQuality: FilterQuality,
-    clipToBounds: Boolean = true,
-    keepContentNoneStartOnDraw: Boolean = false,
-) {
-    val request = requestOfWithSizeResolver(
-        model = state.model,
-        contentScale = contentScale,
-    )
-    validateRequest(request)
-
-    Layout(
-        modifier = modifier.then(
-            ContentPainterElement(
-                request = request,
-                imageLoader = state.imageLoader,
-                modelEqualityDelegate = state.modelEqualityDelegate,
-                transform = transform,
-                onState = onState,
-                contentScale = contentScale,
-                filterQuality = filterQuality,
-                alignment = alignment,
-                alpha = alpha,
-                colorFilter = colorFilter,
-                clipToBounds = clipToBounds,
-                keepContentNoneStartOnDraw = keepContentNoneStartOnDraw,
-                previewHandler = previewHandler(),
-                contentDescription = contentDescription,
-            ),
-        ),
-        measurePolicy = UseMinConstraintsMeasurePolicy,
-    )
-}
-
 private fun onState(
     context: PlatformContext,
     coroutineScope: CoroutineScope,
@@ -353,13 +271,19 @@ private fun onState(
 ) {
     val finaData = if (model is ImageRequest) model.data else model
     zoomState.zoomable.logger.d {
-        "CoilZoomAsyncImage. onState. state=${loadState.name}. data='${finaData}'"
+        val stateName = when (loadState) {
+            is State.Loading -> "Loading"
+            is State.Success -> "Success"
+            is State.Error -> "Error"
+            is State.Empty -> "Empty"
+        }
+        "CoilZoomAsyncImage. onState. state=$stateName. data='${finaData}'"
     }
     val painterSize = loadState.painter
         ?.intrinsicSize
         ?.takeIf { it.isSpecified }
-        ?.roundToIntSize()
-        ?.takeIf { it.isNotEmpty() }
+        ?.let { IntSize(it.width.roundToInt(), it.height.roundToInt()) }
+        ?.takeIf { it.width > 0 && it.height > 0 }
     zoomState.zoomable.contentSize = painterSize ?: IntSize.Zero
 
     if (loadState is State.Success) {
@@ -382,17 +306,3 @@ private fun onState(
         zoomState.setSubsamplingImage(null as SubsamplingImage?)
     }
 }
-
-private val State.name: String
-    get() = when (this) {
-        is State.Loading -> "Loading"
-        is State.Success -> "Success"
-        is State.Error -> "Error"
-        is State.Empty -> "Empty"
-    }
-
-private fun Size.roundToIntSize(): IntSize {
-    return IntSize(width.roundToInt(), height.roundToInt())
-}
-
-private fun IntSize.isNotEmpty(): Boolean = width > 0 && height > 0