+++ zoomimage-compose-coil3-core/src/commonMain/kotlin/com/github/panpf/zoomimage/compose/coil/CoilComposeSubsamplingImageGenerator.kt
@@ -16,7 +16,6 @@
 
 package com.github.panpf.zoomimage.compose.coil
 
-import com.github.panpf.zoomimage.subsampling.ImageSource as ZoomImageImageSource
 import androidx.compose.ui.graphics.painter.Painter
 import coil3.ImageLoader
 import coil3.PlatformContext
@@ -24,7 +23,7 @@ import coil3.request.SuccessResult
 import com.github.panpf.zoomimage.subsampling.SubsamplingImageGenerateResult
 
 /**
- * Convert coil model to [ZoomImageImageSource.Factory]
+ * Convert coil SuccessResult to [SubsamplingImageGenerateResult]
  */
 interface CoilComposeSubsamplingImageGenerator {